+++ main.py
@@ -1,5 +1,7 @@
+import sys
 from sklearn.datasets import load_iris
 from sklearn.model_selection import train_test_split
+from forest import test_forest, train_forest
 from tree import train_tree
 
 data_iris = load_iris()
@@ -16,5 +18,15 @@ def main():
     )
 
 
+def forest():
+    forest = train_forest(train, train_output, 50, len(train) - 10)
+    print(
+        f"Trained on {len(train)} Data.\nTested on {len(test)} Data.\nAccuracy: {round(test_forest(forest, test, test_output), 2)}%"
+    )
+
+
 if __name__ == "__main__":
-    main()
+    if len(sys.argv) > 1 and sys.argv[1] == "forest":
+        forest()
+    else:
+        main()